embedding/requirements.txt
cohere==5.0.2
qdrant-client==1.7.0
python-dotenv==1.0.0